Extract pdf url generate to a custom hook
