+++ libs/application/templates/family-matters/core/src/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useGeneratePdfUrl } from './useGeneratePdfUrl'
